nixd/completion/option: split options into nix extension library
