+++ lib/nixd/include/nixd/nix/Value.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <nix/attr-path.hh>
+#include <nix/eval.hh>
+
+namespace nix::nixd {
+
+bool isOption(EvalState &State, Value &V);
+
+std::optional<std::string> attrPathStr(nix::EvalState &State, nix::Value &V,
+                                       const std::string &AttrPath) noexcept;
+
+} // namespace nix::nixd
